# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth API - Login & Protect",
    "version": "1.0.0",
    "description": "Supabase Auth with JWT Bearer Protection"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "test@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login and receive JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "test@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "JWT returned" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Open public endpoint",
        "responses": { "200": { "description": "Public info message" } }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get authenticated user profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User details" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Get authenticated user dashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard data" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out successfully" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}